quic util cleanup
