+++ quic-forward-proxy/src/lib.rs
@@ -11,7 +11,6 @@ mod cli;
 mod test_client;
 mod util;
 mod tx_store;
-mod tpu_quic_connection_utils;
 mod quinn_auto_reconnect;
 mod outbound;
 mod inbound;
