+++ quic-forward-proxy/src/tls_config_provider_server.rs
@@ -1,8 +1,4 @@
-use std::sync::atomic::{AtomicU32, Ordering};
-use rcgen::generate_simple_self_signed;
-use rustls::{Certificate, ClientConfig, PrivateKey, ServerConfig};
-use crate::tpu_quic_connection_utils::SkipServerVerification;
-use crate::quic_util::ALPN_TPU_FORWARDPROXY_PROTOCOL_ID;
+use rustls::ServerConfig;
 
 // TODO integrate with tpu_service + quic_connection_utils
