+++ quic-forward-proxy/src/tls_self_signed_pair_generator.rs
@@ -1,8 +1,7 @@
 use std::sync::atomic::{AtomicU32, Ordering};
 use rcgen::generate_simple_self_signed;
 use rustls::{Certificate, ClientConfig, PrivateKey, ServerConfig};
-use crate::tpu_quic_connection_utils::SkipServerVerification;
-use crate::quic_util::ALPN_TPU_FORWARDPROXY_PROTOCOL_ID;
+use crate::quic_util::{ALPN_TPU_FORWARDPROXY_PROTOCOL_ID, SkipServerVerification};
 use crate::tls_config_provider_client::TpuCLientTlsConfigProvider;
 use crate::tls_config_provider_server::ProxyTlsConfigProvider;
